+++ Engine/src/Platform/Vulkan/Internal/VulkanBuffer.cpp
@@ -1,7 +1,7 @@
 #include "Platform/Vulkan/Internal/VulkanBuffer.h"
 
 #include "Platform/Vulkan/Internal/VulkanUtils.h"
-#include "Platform/Vulkan/Internal/VulkanCore.h"
+#include "Platform/Vulkan/Internal/VulkanWrappers.h"
 
 namespace aero3d {
