+++ Engine/src/Platform/Vulkan/VulkanAllocators.h
@@ -3,7 +3,6 @@
 
 #include <vector>
 #include <list>
-#include <mutex>
 #include <stdexcept>
 #include <algorithm>
